Replaces use of transformatons.map with custom Live Data class
Using a custom class lets us move the transformation work off the main thread, improving rendering times.
