+++ app/src/main/java/gov/wa/wsdot/android/wsdot/di/FragmentModule.java
@@ -11,7 +11,7 @@ import gov.wa.wsdot.android.wsdot.ui.camera.CameraImageFragment;
 import gov.wa.wsdot.android.wsdot.ui.ferries.bulletins.FerriesRouteAlertsBulletinDetailsFragment;
 import gov.wa.wsdot.android.wsdot.ui.ferries.bulletins.FerriesRouteAlertsBulletinsFragment;
 import gov.wa.wsdot.android.wsdot.ui.ferries.departures.FerriesRouteSchedulesDayDeparturesFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.departures.FerriesTerminalCameraFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.cameras.FerriesTerminalCameraFragment;
 import gov.wa.wsdot.android.wsdot.ui.ferries.FerriesRouteSchedulesFragment;
 import gov.wa.wsdot.android.wsdot.ui.ferries.departures.vesselwatch.VesselWatchFragment;
 import gov.wa.wsdot.android.wsdot.ui.home.DashboardFragment;
