+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/departures/cameras/FerriesTerminalCameraFragment.java
@@ -16,7 +16,7 @@
  *
  */
 
-package gov.wa.wsdot.android.wsdot.ui.ferries.departures;
+package gov.wa.wsdot.android.wsdot.ui.ferries.departures.cameras;
 
 import android.arch.lifecycle.ViewModelProvider;
 import android.arch.lifecycle.ViewModelProviders;
