+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/departures/cameras/FerryTerminalCameraViewModel.java
@@ -1,4 +1,4 @@
-package gov.wa.wsdot.android.wsdot.ui.ferries.departures;
+package gov.wa.wsdot.android.wsdot.ui.ferries.departures.cameras;
 
 import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.MediatorLiveData;
@@ -6,18 +6,16 @@ import android.arch.lifecycle.MutableLiveData;
 import android.arch.lifecycle.ViewModel;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
 
 import javax.inject.Inject;
 
 import gov.wa.wsdot.android.wsdot.database.cameras.CameraEntity;
 import gov.wa.wsdot.android.wsdot.repository.CameraRepository;
 import gov.wa.wsdot.android.wsdot.shared.CameraItem;
-import gov.wa.wsdot.android.wsdot.shared.FerriesTerminalItem;
 import gov.wa.wsdot.android.wsdot.ui.camera.MapCameraViewModel;
+import gov.wa.wsdot.android.wsdot.ui.ferries.helpers.FerryHelper;
 import gov.wa.wsdot.android.wsdot.util.AppExecutors;
 import gov.wa.wsdot.android.wsdot.util.network.ResourceStatus;
 
@@ -28,8 +26,6 @@ public class FerryTerminalCameraViewModel extends ViewModel {
     private MutableLiveData<ResourceStatus> mStatus;
 
     private MediatorLiveData<List<CameraItem>> terminalCameras;
-
-    private static Map<Integer, FerriesTerminalItem> ferriesTerminalMap = new HashMap<>();
 
     private AppExecutors appExecutors;
     private CameraRepository cameraRepo;
@@ -56,25 +52,29 @@ public class FerryTerminalCameraViewModel extends ViewModel {
 
     public void processTerminalCameras(Integer terminalId, List<CameraEntity> cameras){
 
-        List<CameraItem> terminalCameraItems = new ArrayList<>();
+        appExecutors.taskIO().execute(() -> {
+            List<CameraItem> terminalCameraItems = new ArrayList<>();
 
-        for (CameraEntity cameraEntity: cameras) {
+            for (CameraEntity cameraEntity: cameras) {
 
-            int distance = FerryHelper.getDistanceFromTerminal(terminalId, cameraEntity.getLatitude(), cameraEntity.getLongitude());
+                int distance = FerryHelper.getDistanceFromTerminal(terminalId, cameraEntity.getLatitude(), cameraEntity.getLongitude());
 
-            // If less than 3 miles from terminal, and labeled as a ferries camera, show it
-            if (distance < 15840 && cameraEntity.getRoadName().toLowerCase(Locale.US).equals("ferries")) { // in feet
-                CameraItem camera = new CameraItem();
-                camera.setCameraId(cameraEntity.getCameraId());
-                camera.setTitle(cameraEntity.getTitle());
-                camera.setImageUrl(cameraEntity.getUrl());
-                camera.setLatitude(cameraEntity.getLatitude());
-                camera.setLongitude(cameraEntity.getLongitude());
-                camera.setDistance(distance);
-                terminalCameraItems.add(camera);
+                // If less than 3 miles from terminal, and labeled as a ferries camera, show it
+                if (distance < 15840 && cameraEntity.getRoadName().toLowerCase(Locale.US).equals("ferries")) { // in feet
+                    CameraItem camera = new CameraItem();
+                    camera.setCameraId(cameraEntity.getCameraId());
+                    camera.setTitle(cameraEntity.getTitle());
+                    camera.setImageUrl(cameraEntity.getUrl());
+                    camera.setLatitude(cameraEntity.getLatitude());
+                    camera.setLongitude(cameraEntity.getLongitude());
+                    camera.setDistance(distance);
+                    terminalCameraItems.add(camera);
+                }
             }
-        }
-        terminalCameras.postValue(terminalCameraItems);
+            terminalCameras.postValue(terminalCameraItems);
+        });
+
+
     }
 
 }
